// src/common/enum/common.enum.ts
/**
 * GENDER
 */
 export enum GENDER {
    MALE = "male",
    FEMALE = "female",
}

/**
 * Status
 */
export enum STATUS_ENUM {
    ACTIVE = "active",
    INACTIVE = "inactive",
}

/**
 * Delete Enum
 */
export enum DELETE_ENUM {
    YES = 1,
    NO = 0
}

/**
 * Mode of APPROVE 
 */
export enum APPROVE_ENUM {
    PENDING = "pending",
    APPROVED = "aprroved",
    REJECTED = "rejected",
}

/**
 * Roles
 */
export enum ROLE_ENUM {
    ADMIN = "admin",
    USER = "user",
    TECHNICIAN = "technician"
}